Moved HttpRequest to utils package. Fixed casing of commands
