+++ src/main/java/de/devgruppe/fundiscordbot/command/commands/GifTrending.java
@@ -7,7 +7,7 @@ import com.google.gson.JsonParser;
 import de.devgruppe.fundiscordbot.FunDiscordBotStarter;
 import de.devgruppe.fundiscordbot.command.Command;
 import de.devgruppe.fundiscordbot.command.CommandResponse;
-import de.devgruppe.fundiscordbot.util.HttpRequest;
+import de.devgruppe.fundiscordbot.utils.HttpRequest;
 import java.io.IOException;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.entities.Message;
