+++ src/main/java/de/devgruppe/fundiscordbot/command/commands/MemeNamesCommand.java
@@ -4,7 +4,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import de.devgruppe.fundiscordbot.command.Command;
 import de.devgruppe.fundiscordbot.command.CommandResponse;
-import de.devgruppe.fundiscordbot.util.HttpRequest;
+import de.devgruppe.fundiscordbot.utils.HttpRequest;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
