+++ src/main/java/de/devgruppe/fundiscordbot/command/impl/DefaultCommandRegistry.java
@@ -3,15 +3,14 @@ package de.devgruppe.fundiscordbot.command.impl;
 import de.devgruppe.fundiscordbot.command.Command;
 import de.devgruppe.fundiscordbot.command.CommandRegistry;
 import de.devgruppe.fundiscordbot.command.CommandResponse;
-import net.dv8tion.jda.core.entities.ChannelType;
-import net.dv8tion.jda.core.entities.Message;
-import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
-import net.dv8tion.jda.core.hooks.ListenerAdapter;
-
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import net.dv8tion.jda.core.entities.ChannelType;
+import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+import net.dv8tion.jda.core.hooks.ListenerAdapter;
 
 /**
  * Created by GalaxyHD on 11.09.2017.
@@ -67,7 +66,7 @@ public class DefaultCommandRegistry extends ListenerAdapter implements CommandRe
   @Override
   public Command getCommandObjectByName(String commandName) {
     return this.commands.stream()
-        .filter(command -> command.getCommandName().equals(commandName))
+        .filter(command -> command.getCommandName().equalsIgnoreCase(commandName))
         .findFirst()
         .orElse(null);
   }
